+++ include/libcyphal/transport/can/delegate.hpp
@@ -128,7 +128,7 @@ public:
     ///
     virtual void acceptRxTransfer(CanardMemory&&            canard_memory,
                                   const TransferRxMetadata& rx_metadata,
-                                  const CanardNodeID        source_node_id) = 0;
+                                  const NodeId              source_node_id) = 0;
 
 protected:
     IRxSessionDelegate()  = default;
@@ -571,7 +571,7 @@ private:
 
         void acceptRxTransfer(CanardMemory&&            canard_memory,
                               const TransferRxMetadata& rx_metadata,
-                              const CanardNodeID        source_node_id) override
+                              const NodeId              source_node_id) override
         {
             // This is where de-multiplexing happens: the transfer is forwarded to the appropriate session.
             // It's ok not to find the session delegate here - we drop unsolicited transfers.
